extracted karma functions
also changed length back to 5 because the weight function reaches higher values
